[FIX]: 로고 업데이트 메소드에서 로고가 이미 삭제 되었을 경우, 삭제 전에 null 체크

+++ src/main/java/studio/studioeye/domain/company_information/application/CompanyInformationService.java
@@ -220,9 +220,9 @@ public class CompanyInformationService {
             return ApiResponse.withError(ErrorCode.COMPANYINFORMATION_IS_EMPTY);
         }
         String fileName = companyInformations.get(0).getLightLogoImageFileName();
-        s3Adapter.deleteFile(fileName);
+        if(fileName != null) s3Adapter.deleteFile(fileName);
         fileName = companyInformations.get(0).getDarkLogoImageFileName();
-        s3Adapter.deleteFile(fileName);
+        if(fileName != null) s3Adapter.deleteFile(fileName);
 
         ApiResponse<String> updateLightLogoFileResponse = s3Adapter.uploadFile(lightLogoImage);
         if (updateLightLogoFileResponse.getStatus().is5xxServerError()) {
